+++ node/clients/index.ts
@@ -1,17 +1,27 @@
 import { IOClients } from '@vtex/api'
-import { masterDataFor, Checkout, OMS } from '@vtex/clients'
-import type { Affiliates } from 'vtex.affiliates'
+import { masterDataFor } from '@vtex/clients'
+import type { Affiliates, UserAffiliation } from 'vtex.affiliates'
+
+import AuthenticationClient from './authenticationClient'
+import CheckoutExtended from './checkout'
 
 export class Clients extends IOClients {
   public get affiliates() {
     return this.getOrSet('affiliates', masterDataFor<Affiliates>('affiliates'))
   }
 
-  public get checkout() {
-    return this.getOrSet('checkout', Checkout)
+  public get userAffiliation() {
+    return this.getOrSet(
+      'userAffiliation',
+      masterDataFor<UserAffiliation>('userAffiliation')
+    )
   }
 
-  public get oms() {
-    return this.getOrSet('oms', OMS)
+  public get checkout() {
+    return this.getOrSet('checkout', CheckoutExtended)
+  }
+
+  public get authentication() {
+    return this.getOrSet('authentication', AuthenticationClient)
   }
 }