+++ node/index.ts
@@ -8,7 +8,6 @@ import { method, Service } from '@vtex/api'
 
 import { Clients } from './clients'
 import { createAffiliate } from './middlewares/createAffiliate'
-import { getClient } from './middlewares/getClient'
 import { getOrder } from './middlewares/getOrder'
 import { setupAppConfiguration } from './middlewares/setupAppConfiguration'
 import { updateAffiliate } from './middlewares/updateAffiliate'
@@ -21,11 +20,12 @@ import { isAffiliateValid } from './resolvers/isAffiliateValid'
 import { setAffiliateOnOrderForm } from './resolvers/setAffiliateOnOrderForm'
 import type { AffiliateInput } from './typings/affiliates'
 import { setAffiliateLeadOnCustomData } from './middlewares/setAffiliateLeadOnCustomData'
-import { verifyOrderAffiliation } from './middlewares/verifyOrderAffiliation'
+import { verifyOrderFormAffiliation } from './middlewares/verifyOrderFormAffiliation'
 import { getOrderForm } from './middlewares/getOrderForm'
-import { getClientFromOrderForm } from './middlewares/getClientFromOrderForm'
-import { verifyClientAffiliation } from './middlewares/verifyClientAffiliation'
 import { getAffiliateStoreName } from './resolvers/getAffiliateStoreName'
+import { getAffiliateLead } from './middlewares/getAffiliateLead'
+import { verifyUserAffiliation } from './middlewares/verifyUserAffiliation'
+import { authenticateRequest } from './middlewares/authenticateRequest'
 
 const TIMEOUT_MS = 1000
 
@@ -74,8 +74,8 @@ export default new Service({
   clients,
   routes: {
     affiliate: method({
-      POST: [validateCreate, createAffiliate],
-      PATCH: [validateUpdate, updateAffiliate],
+      POST: [authenticateRequest, validateCreate, createAffiliate],
+      PATCH: [authenticateRequest, validateUpdate, updateAffiliate],
     }),
   },
   graphql: {
@@ -94,15 +94,15 @@ export default new Service({
     setAffiliateLead: [
       getOrder,
       validateCustomData,
-      getClient,
+      getAffiliateLead,
       validateLead,
       updateLead,
     ],
     verifyUserAffiliateLead: [
       getOrderForm,
-      verifyOrderAffiliation,
-      getClientFromOrderForm,
-      verifyClientAffiliation,
+      verifyOrderFormAffiliation,
+      getAffiliateLead,
+      verifyUserAffiliation,
       setAffiliateLeadOnCustomData,
     ],
   },